requirements.txt
# RealSense Camera 相关依赖
pyrealsense2>=2.50.0
opencv-python>=4.5.0
numpy>=1.19.0

# YOLO 和 SAM 模型依赖
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0

# 图像处理
pillow>=8.0.0
matplotlib>=3.3.0

